refactor: move gremlin config into config folder
