+++ src/config/gremlins/index.ts
@@ -1,0 +1,1 @@
+export { gremlins } from './gremlins';
