+++ src/config/index.ts
@@ -2,3 +2,4 @@ import { GameActionId as GameActionIdT } from './rounds';
 export type GameActionId = GameActionIdT;
 
 export { rounds } from './rounds';
+export { gremlins } from './gremlins';
